requirement.txt
openai-whisper
pypinyin
pykakasi